Template update

+++ Kros.ProjectTemplates/src/Kros.CqrsTemplate/content/Application/Commands/CreateRRREntityNameRRR_/CreateRRREntityNameRRR_Command.cs
@@ -5,7 +5,7 @@ namespace Kros.CqrsTemplate.Application.Commands
     /// <summary>
     /// Create RRREntityNameRRR_ command.
     /// </summary>
-    public class CreateRRREntityNameRRR_Command : IRequest<int>
+    public class CreateRRREntityNameRRR_Command : IRequest<long>
     {
     }
 }
